+++ contracts/Cargo.toml
@@ -1,15 +1,20 @@
-[package]
-name = "vertexchain-contracts"
+[workspace]
+resolver = "2"
+members = [
+    "gist-registry",
+    "multisig",
+    "governance",
+    "batch-wallet",
+]
+
+[workspace.package]
 version = "0.1.0"
 edition = "2021"
 
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
+[workspace.dependencies]
 soroban-sdk = { version = "22.0.0", features = ["alloc"] }
 
-[dev-dependencies]
+[workspace.dev-dependencies]
 soroban-sdk = { version = "22.0.0", features = ["testutils"] }
 
 [profile.release]
