+++ contracts/README.md
@@ -22,22 +22,25 @@ The contracts handle registering gists as verifiable blockchain records, organiz
 
 ```
 contracts/
-├── src/
-│   └── lib.rs       # GistRegistry contract
-├── Cargo.toml
+├── gist-registry/       # Gist registry contract
+├── multisig/            # Multi-signature wallet contract
+├── governance/          # Governance and proposal contract
+├── batch-wallet/        # Batch wallet creation/recovery contract
+├── Cargo.toml           # Workspace configuration
 └── README.md
 ```
 
-All contracts live in a single crate for now, with flexibility to split into separate packages as complexity grows.
+This is a Cargo workspace with separate packages for each contract, allowing for independent development and deployment.
 
 ---
 
-## GistRegistry Contract (MVP)
+## Available Contracts
 
-### Data Model
+### GistRegistry (MVP)
 
-Each gist record tracks:
+Location-aware gist registry for storing and retrieving gists by geographic cells.
 
+**Data Model:**
 | Field | Type | Description |
 |---|---|---|
 | `gist_id` | `u64` | Auto-incremented identifier |
@@ -46,23 +49,60 @@ Each gist record tracks:
 | `content_hash` | `String` | Content hash pointer (e.g. IPFS CID) |
 | `created_at` | `u64` | Ledger timestamp at creation |
 
-### Public Methods
+**Methods:**
+- `post_gist(author, location_cell, content_hash)` - Register a new gist
+- `get_gist(gist_id)` - Retrieve a gist record by id
+- `list_gists_by_cell(location_cell, cursor, limit)` - Paginated list of gists within a location cell
 
-| Method | Description |
-|---|---|
-| `post_gist(author, location_cell, content_hash)` | Register a new gist; returns its `gist_id` |
-| `get_gist(gist_id)` | Retrieve a gist record by id |
-| `list_gists_by_cell(location_cell, cursor, limit)` | Paginated list of gists within a location cell |
+### Multisig
 
----
+Multi-signature wallet contract for requiring multiple approvals before transactions can be executed.
 
-## Planned Contracts
+**Features:**
+- Configurable signers and threshold
+- Transaction submission and approval tracking
+- High-value transaction thresholds
+- Event emission for all operations
 
-| Contract | Purpose |
-|---|---|
-| **Tipping** | Tip mechanisms for gist authors |
-| **Staking** | Stakeholder systems |
-| **Moderation** | On-chain content moderation |
+**Methods:**
+- `initialize(admin)` - Initialize the multisig contract
+- `set_signers(caller, signers, threshold)` - Configure signers and approval threshold
+- `submit_transaction(caller, to, amount, payload, asset)` - Submit a transaction for approval
+- `approve_transaction(caller, tx_id)` - Approve a pending transaction
+- `execute_transaction(caller, tx_id)` - Execute an approved transaction
+
+### Governance
+
+Governance contract for proposing and voting on configuration changes.
+
+**Features:**
+- Proposal creation with configurable duration
+- Voting mechanism with approval thresholds
+- Configurable string-based key-value storage
+- Admin management
+
+**Methods:**
+- `initialize(admin, required_approvals)` - Initialize governance
+- `create_proposal(proposer, config_key, config_value, duration_seconds)` - Create a proposal
+- `vote_proposal(voter, proposal_id)` - Vote on a proposal
+- `execute_proposal(caller, proposal_id)` - Execute an approved proposal
+- `get_config(config_key)` - Retrieve a configuration value
+
+### BatchWallet
+
+Batch wallet creation and recovery contract for efficient wallet management.
+
+**Features:**
+- Batch wallet creation with duplicate detection
+- Batch wallet recovery operations
+- Event emission for batch operations
+- Admin-controlled operations
+
+**Methods:**
+- `initialize(admin)` - Initialize the batch wallet contract
+- `batch_create_wallets(caller, requests)` - Create multiple wallets in a single transaction
+- `batch_recover_wallets(caller, requests)` - Recover multiple wallets in a single transaction
+- `get_wallet(owner)` - Retrieve wallet information
 
 ---
 
@@ -86,23 +126,38 @@ rustup target add wasm32-unknown-unknown
 cargo install --locked stellar-cli --features opt
 ```
 
-### Build
+### Build all contracts
 
 ```bash
 cargo build --target wasm32-unknown-unknown --release
 ```
 
-### Test
+### Build specific contract
+
+```bash
+cd gist-registry
+cargo build --target wasm32-unknown-unknown --release
+```
+
+### Test all contracts
 
 ```bash
 cargo test
 ```
 
-### Deploy (local testnet)
+### Test specific contract
 
 ```bash
+cd gist-registry
+cargo test
+```
+
+### Deploy specific contract (local testnet)
+
+```bash
+cd gist-registry
 stellar contract deploy \
-  --wasm target/wasm32-unknown-unknown/release/vertexchain_contracts.wasm \
+  --wasm target/wasm32-unknown-unknown/release/gist_registry.wasm \
   --network testnet \
   --source <your-identity>
 ```
@@ -114,6 +169,7 @@ stellar contract deploy \
 - Modifications to contract interfaces require prior discussion via a linked issue and design documentation.
 - Public functions should remain compact and well-documented.
 - New functionality must be accompanied by test coverage.
+- When adding new contracts, create a new package in the workspace.
 
 ---
