+++ contracts/batch-wallet/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "batch-wallet"
+version.workspace = true
+edition.workspace = true
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+soroban-sdk.workspace = true
+
+[dev-dependencies]
+soroban-sdk.workspace = true
