+++ contracts/gist-registry/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "gist-registry"
+version.workspace = true
+edition.workspace = true
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+soroban-sdk.workspace = true
+
+[dev-dependencies]
+soroban-sdk.workspace = true
