+++ contracts/gist-registry/src/lib.rs
