+++ contracts/governance/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "governance"
+version.workspace = true
+edition.workspace = true
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+soroban-sdk.workspace = true
+
+[dev-dependencies]
+soroban-sdk.workspace = true
